Code refactor KT-2
