+++ package.json
@@ -30,6 +30,7 @@
         "typescript": "^4.2.4"
     },
     "prettier": {
+        "jsxSingleQuote":true,
         "singleQuote": true,
         "trailingComma": "all",
         "printWidth": 120,
