+++ src/api/dtos/currency.dto.ts
@@ -1,9 +1,8 @@
-import {CurrencyIso} from 'core/utils/currencyList';
 import {ExchangeRates} from '../../core/models';
 
 export interface CurrencyPairsDto {
     /** Текущая базовая валюта. */
-    base: CurrencyIso;
+    base: string;
 
     /** Дата запроса. */
     date: string;
